fix: No zippping and deleting when list_Only_Folders is True

+++ Mail_downloader.py
@@ -11,7 +11,7 @@ import time
 Path_config = ""
 if len(sys.argv) >= 2:
     Arg_SingleDB = sys.argv[1]
-    Path_config =f"config/{Arg_SingleDB.replace("--config_file=", "")}"
+    Path_config = f"config/{Arg_SingleDB.replace("--config_file=", "")}"
 else:
     Path_config = "config/config.ini"
 
@@ -121,16 +121,16 @@ with MailBox(imap_server, port=imap_port).login_utf8(
                     errorcounter += 1
                     continue
 
+if not list_Only_Folders:
+    zipfolder(f"{imap_server}_{date}", f"export/{imap_server}")
+    shutil.rmtree(f"export/{imap_server}")
 
-zipfolder(f"{imap_server}_{date}", f"export/{imap_server}")
-shutil.rmtree(f"export/{imap_server}")
+    if days_to_delete > 0:
+        for filename in os.listdir(ZIP_export_folder):
+            filepath = os.path.join(ZIP_export_folder, filename)
 
-if days_to_delete > 0:
-    for filename in os.listdir(ZIP_export_folder):
-        filepath = os.path.join(ZIP_export_folder, filename)
+            if os.path.isfile(filepath):
+                creation_time = os.path.getmtime(filepath)
 
-        if os.path.isfile(filepath):
-            creation_time = os.path.getmtime(filepath)
-
-            if (now - creation_time) > days_to_delete:
-                os.remove(filepath)
+                if (now - creation_time) > days_to_delete:
+                    os.remove(filepath)
